refactor(tests): move shared mount-safe helpers into _mount_safe_helpers
test_server_fs_raw_mount_safe imported fixtures/helpers from the
test_server_fs_mount_safe module (one test file reaching into another). Move
the shared `home` fixture and the _mount/_entry/_no_kernel_on_mount/
_list_returns/_list_raises helpers into a neutral tests/_mount_safe_helpers
module both suites import from. No test behavior change.
